+++ src/main/kotlin/com/krillsson/sysapi/core/monitoring/Extensions.kt
@@ -59,6 +59,7 @@ fun <E : Enum<E>> Monitor.Type.toEnumEntries(): EnumEntries<E>? =
         Monitor.Type.PROCESS_EXISTS,
         Monitor.Type.CONNECTIVITY,
         Monitor.Type.WEBSERVER_UP,
+        Monitor.Type.CHECK_LATENCY,
         Monitor.Type.EXTERNAL_IP_CHANGED,
         Monitor.Type.UPS_OPERATING_NORMALLY,
         Monitor.Type.UPS_LOAD_WATT,