+++ src/main/kotlin/com/krillsson/sysapi/notifications/localization/OngoingEventFormatter.kt
@@ -36,6 +36,7 @@ class OngoingEventFormatter(
                 Monitor.Type.PROCESS_EXISTS -> "Process stopped on $serverName"
                 Monitor.Type.CONNECTIVITY -> "Connection is down on $serverName"
                 Monitor.Type.WEBSERVER_UP -> "Check is failing on $serverName"
+                Monitor.Type.CHECK_LATENCY -> "Check is slow on $serverName"
                 Monitor.Type.EXTERNAL_IP_CHANGED -> "External IP changed on $serverName"
                 Monitor.Type.UPS_LOAD_PERCENTAGE -> "Load too high on UPS $serverName"
                 Monitor.Type.UPS_OPERATING_NORMALLY -> "UPS not operating normally on $serverName"
@@ -98,6 +99,7 @@ class OngoingEventFormatter(
                 } CPU usage went above $formattedThreshold to $formattedValue"
 
                 Monitor.Type.WEBSERVER_UP -> "$monitoredItemId is failing its check"
+                Monitor.Type.CHECK_LATENCY -> "$monitoredItemId answered in $formattedValue, above $formattedThreshold"
                 Monitor.Type.DISK_TEMPERATURE -> "Temperature on $monitoredItemId went above $formattedThreshold to $formattedValue"
                 Monitor.Type.UPS_OPERATING_NORMALLY -> "UPS $monitoredItemId is not operating normally"
                 Monitor.Type.UPS_LOAD_PERCENTAGE -> "UPS $monitoredItemId load went above $formattedThreshold to $formattedValue"